move navbar logged in/out back to navbar.js
