+++ src/components/login/NavBar.js
@@ -1,17 +1,15 @@
 import { useNavigate, Link } from 'react-router-dom';
-import NavLoggedIn from './NavLoggedIn';
-import NavLoggedOut from './NavLoggedOut';
 import Search from '../Search';
 
 function NavBar({ token, setToken, user, JWT }) {
-	// let navigate = useNavigate();
+	let navigate = useNavigate();
 
-	// function handleLogOut(e) {
-	// 	e.preventDefault();
-	// 	setToken(false); //can be null as well?
-	// 	localStorage.clear();
-	// 	navigate('/');
-	// }
+	function handleLogOut(e) {
+		e.preventDefault();
+		setToken(false); //can be null as well?
+		localStorage.clear();
+		navigate('/');
+	}
 
 	return (
 		<nav>
@@ -22,12 +20,17 @@ function NavBar({ token, setToken, user, JWT }) {
 				{/* Requires logged in status */}
 				{token ? (
 					<div>
-						<NavLoggedIn setToken={setToken} />
 						<Link to='/user-rec-form'>Add a Recommendation</Link>
+						<p></p>
+						<Link to='/user-recs'>Your Page</Link>
+						<p></p>
+						<Link to='' onClick={handleLogOut}>
+							Log Out
+						</Link>
 					</div>
 				) : (
 					// Show below if not logged in
-					<NavLoggedOut />
+					<Link to='/auth'>Log In</Link>
 				)}
 			</div>
 		</nav>
